+++ app/src/main/AndroidManifest.xml
@@ -25,27 +25,27 @@
             android:name="firebase_crashlytics_collection_enabled"
             android:value="true" />
 
-        <activity android:name=".LoginActivity">
+        <activity android:name=".ui.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:windowSoftInputMode="adjustPan"
             android:hardwareAccelerated="true">
         </activity>
 
-        <activity android:name=".BuscarFinActivity" android:label="Buscar Fin"/>
-        <activity android:name=".EditFinActivity" android:label="Edit Fin"/>
-        <activity android:name=".NewFinActivity" android:label="Nova Fin"/>
-        <activity android:name=".ResultBuscaActivity" android:label="Resumo Fin"/>
-        <activity android:name=".ResumoDespGrafActivity" android:exported="false" />
-        <activity android:name=".ResultBuscaCredActivity" android:label="Resumo Créditos Fin"/>
-        <activity android:name=".ResultBuscaGrafActivity" android:label="Resumo Graf Fin"/>
-        <activity android:name=".ResultBuscaDespActivity" android:label="Resumo Despesas Fin"/>
+        <activity android:name=".ui.BuscarFinActivity" android:label="Buscar Fin"/>
+        <activity android:name=".ui.EditFinActivity" android:label="Edit Fin"/>
+        <activity android:name=".ui.NewFinActivity" android:label="Nova Fin"/>
+        <activity android:name=".ui.ResultBuscaActivity" android:label="Resumo Fin"/>
+        <activity android:name=".ui.ResumoDespGrafActivity" android:exported="false" />
+        <activity android:name=".ui.ResultBuscaCredActivity" android:label="Resumo Créditos Fin"/>
+        <activity android:name=".ui.ResultBuscaGrafActivity" android:label="Resumo Graf Fin"/>
+        <activity android:name=".ui.ResultBuscaDespActivity" android:label="Resumo Despesas Fin"/>
 
     </application>
 </manifest>
