+++ app/src/main/java/com/app/fintoday/data/FinRepository.java
@@ -1,4 +1,4 @@
-package com.app.fintoday;
+package com.app.fintoday.data;
 
 import android.app.Application;
 import android.util.Log;
