+++ app/src/main/java/com/app/fintoday/ui/LoginActivity.java
@@ -1,13 +1,13 @@
-package com.app.fintoday;
+package com.app.fintoday.ui;
 //Criado em 16.05.25
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.widget.Button;
 import android.widget.Toast;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.app.fintoday.R;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.auth.FirebaseAuth;
