+++ app/src/main/res/layout/activity_busca_fin.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BuscarFinActivity">
+    tools:context=".ui.BuscarFinActivity">
 
     <!-- LinearLayout para o conteúdo existente -->
     <LinearLayout
