+++ app/src/main/res/layout/activity_new_fin.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".NewFinActivity">
+    tools:context=".ui.NewFinActivity">
 
     <EditText
         android:id="@+id/idEdtValorDesp"
