+++ app/src/main/res/layout/activity_result_busca_graf.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ResultBuscaGrafActivity">
+    tools:context=".ui.ResultBuscaGrafActivity">
 
     <!-- Cabeçalho com título e total -->
     <LinearLayout
