+++ app/src/main/res/layout/activity_result_filtrado.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ResultBuscaCredActivity"> <!-- Pode ser genérico -->
+    tools:context=".ui.ResultBuscaCredActivity"> <!-- Pode ser genérico -->
 
     <!-- Total -->
     <TextView
