+++ app/src/main/res/layout/activity_resumo_desp_graf.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".ResumoDespGrafActivity">
+    tools:context=".ui.ResumoDespGrafActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
